+++ Dynamic/Components/AppleInterfaceStyle+Coordinator.swift
@@ -16,12 +16,21 @@ extension AppleInterfaceStyle {
 public class AppleInterfaceStyleCoordinator: NSObject {
     fileprivate override init() { super.init() }
     
+    private var appearanceObservation: NSKeyValueObservation? {
+        didSet {
+            oldValue?.invalidate()
+        }
+    }
+    
     @objc public func toggleInterfaceStyle() {
         AppleInterfaceStyle.toggle()
     }
     
     public func setup() {
         tearDown()
+        appearanceObservation = NSApp.observe(\.effectiveAppearance) { _, _ in
+            AppleInterfaceStyle.updateWallpaper()
+        }
         guard preferences.scheduled else {
             guard preferences.adjustForBrightness else { return }
             // No need for scheduler, only enable brightness observer
@@ -31,9 +40,12 @@ public class AppleInterfaceStyleCoordinator: NSObject {
         Scheduler.shared.schedule(startBrightnessObserverOnFailure: true)
     }
     
-    public func tearDown() {
+    public func tearDown(stopAppearanceObservation: Bool = true) {
         Scheduler.shared.cancel()
         Connectivity.default.stopObserving()
         ScreenBrightnessObserver.shared.stopObserving()
+        if stopAppearanceObservation {
+            appearanceObservation = nil
+        }
     }
 }